+++ src/hooks/useCirculationLog.js
@@ -16,6 +16,8 @@ export const useCirculationLog = (isLoadingRightAway, queryLoadRecords, loadReco
   const queryParams = queryString.parse(location.search);
   const resultsPaneTitleRef = useRef();
 
+  console.log('queryParams', queryParams)
+
   const defaultSearchParams = {
     queryParams,
     limit: pagination.limit,
@@ -32,10 +34,13 @@ export const useCirculationLog = (isLoadingRightAway, queryLoadRecords, loadReco
 
     setIsLoading(true);
 
-    return queryLoadRecords(defaultSearchParams.offset, hasToCallAPI).then(recordsResponse => {
-      if (!offset) {
-        setRecordsCount(recordsResponse?.totalRecords);
+    // to get the actual totalRecords count, the limit should be set to 0
+    queryLoadRecords(defaultSearchParams.offset, 0).then(totalRecordsResponse => {
+      setRecordsCount(totalRecordsResponse?.totalRecords);
+    });
 
+    return queryLoadRecords(defaultSearchParams.offset).then(recordsResponse => {
+      if (!offset) {
         if (recordsResponse?.totalRecords != null) {
           // eslint-disable-next-line no-unused-expressions
           resultsPaneTitleRef?.current?.focus();