+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [6.0.1] In progress
 
 * Bump `@folio/stripes-acq-components` to 7.0.0. Refs UICIRCLOG-175.
+* Make a separate call with a zero limit to get actual totalRecords. Refs UICIRCLOG-179.
 
 ## 6.0.0 (https://github.com/folio-org/ui-circulation-log/tree/v6.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-circulation-log/compare/v5.0.1...v6.0.0)